Фронт. Недопилен

+++ BastardFat.ThirdVersion.BusinessLogic/Services/Implementation/WorkPlaceServiceImpl.cs
@@ -1,6 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Data.Entity;
+using System.Linq;
 using System.Threading.Tasks;
+using System.Web;
 using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using BastardFat.ThirdVersion.BusinessLogic.Services.Interface;
@@ -14,13 +16,15 @@ namespace BastardFat.ThirdVersion.BusinessLogic.Services.Implementation
     public class WorkPlaceServiceImpl : IWorkPlaceService
     {
         private readonly IWorkPlacesRepository _workPlacesRepository;
+        private readonly IPeoplesRepository _peoplesRepository;
         private readonly IMainUnitOfWork _unitOfWork;
         private readonly IMapper _mapper;
 
-        public WorkPlaceServiceImpl(IWorkPlacesRepository workPlacesRepository, IMainUnitOfWork unitOfWork)
+        public WorkPlaceServiceImpl(IWorkPlacesRepository workPlacesRepository, IMainUnitOfWork unitOfWork, IPeoplesRepository peoplesRepository)
         {
             _workPlacesRepository = workPlacesRepository;
             _unitOfWork = unitOfWork;
+            _peoplesRepository = peoplesRepository;
 
             _mapper = new MapperConfiguration(x =>
             {
@@ -29,9 +33,6 @@ namespace BastardFat.ThirdVersion.BusinessLogic.Services.Implementation
                 x.CreateMap<People, PeopleModel>()
                     .ReverseMap();
             }).CreateMapper();
-
-            
-
         }
 
         public async Task<IEnumerable<WorkPlaceModel>> GetAllWorkPlaces()
@@ -44,34 +45,35 @@ namespace BastardFat.ThirdVersion.BusinessLogic.Services.Implementation
 
         public async Task<WorkPlaceModel> AddWorkPlace(WorkPlaceModel people)
         {
-            var result = _mapper.Map<WorkPlaceModel>(
-                _workPlacesRepository.Add(
-                    _mapper.Map<WorkPlace>(people)
-                )
+            var result = _workPlacesRepository.Add(
+                _mapper.Map<WorkPlace>(people)
             );
             await _unitOfWork.CommitAsync();
-            return result;
+            return _mapper.Map<WorkPlaceModel>(result);
         }
 
         public async Task<WorkPlaceModel> UpdateWorkPlace(WorkPlaceModel people)
         {
-            var result = _mapper.Map<WorkPlaceModel>(
-                _workPlacesRepository.Update(
-                    _mapper.Map<WorkPlace>(people)
-                )
+            if (people.Id == 1) throw new HttpException(406, "Not Acceptable");
+            var result = _workPlacesRepository.Update(
+                _mapper.Map<WorkPlace>(people)
             );
             await _unitOfWork.CommitAsync();
-            return result;
+            return _mapper.Map<WorkPlaceModel>(result);
         }
 
         public async Task<WorkPlaceModel> DeleteWorkPlace(int id)
         {
-            var result = _mapper.Map<WorkPlaceModel>(
-                await _workPlacesRepository.DeleteAsync(id)
-            );
+            if (id == 1) throw new HttpException(406, "Not Acceptable");
+
+            await _peoplesRepository
+                .Query()
+                .Where(p => p.WorkPlaceId == id)
+                .ForEachAsync(p => p.WorkPlaceId = 1);
+
+            var result = await _workPlacesRepository.DeleteAsync(id);
             await _unitOfWork.CommitAsync();
-            return result;
+            return _mapper.Map<WorkPlaceModel>(result);
         }
     }
-
 }
